fixed the backend for the reverbDeals page. it now has a working controller
