+++ Backend/controllers/scoopDealsCtrl.js
@@ -1,0 +1,4 @@
+
+const { priceGuideHelper } = require("../helpers/");
+const { loginAuth, getSingleProduct, getProductsList, getCompShopData } = require("../models/");
+
