+++ Backend/routes/Reverb-Scoopers/myFeedRoute.js
@@ -1,12 +1,6 @@
 const feedRouter = require('express').Router();
 const { assessMyFeedPrices } = require('../../controllers/myFeedCtrl');
 
-
-
-// DELETE THIS
-const { loginAuth, getProductsList, getSingleProduct } = require("../../models/");
-
-
 // "/myFeed"
 feedRouter.get('/', assessMyFeedPrices);
